Split AI chat UI state
