+++ docs/ios-localization.md
@@ -204,7 +204,8 @@ When adding a new language, review these support-heavy areas explicitly:
 - [apps/ios/Flashcards/Flashcards/TransientBannerSupport.swift](/Users/kirill/_my_local/code-local/personal-workspace/flashcards-open-source-app/apps/ios/Flashcards/Flashcards/TransientBannerSupport.swift)
 - [apps/ios/Flashcards/Flashcards/ErrorMessageSupport.swift](/Users/kirill/_my_local/code-local/personal-workspace/flashcards-open-source-app/apps/ios/Flashcards/Flashcards/ErrorMessageSupport.swift)
 - [apps/ios/Flashcards/Flashcards/Cloud/CloudAuthService.swift](/Users/kirill/_my_local/code-local/personal-workspace/flashcards-open-source-app/apps/ios/Flashcards/Flashcards/Cloud/CloudAuthService.swift)
-- [apps/ios/Flashcards/Flashcards/AI/AIChatStore+UIState.swift](/Users/kirill/_my_local/code-local/personal-workspace/flashcards-open-source-app/apps/ios/Flashcards/Flashcards/AI/AIChatStore+UIState.swift)
+- [apps/ios/Flashcards/Flashcards/AI/AIChatStore+AlertPresentation.swift](/Users/kirill/_my_local/code-local/personal-workspace/flashcards-open-source-app/apps/ios/Flashcards/Flashcards/AI/AIChatStore+AlertPresentation.swift)
+- [apps/ios/Flashcards/Flashcards/AI/AIChatStore+ObservabilityCapture.swift](/Users/kirill/_my_local/code-local/personal-workspace/flashcards-open-source-app/apps/ios/Flashcards/Flashcards/AI/AIChatStore+ObservabilityCapture.swift)
 - [apps/ios/Flashcards/Flashcards/AI/AIChatStore+RunOrchestration.swift](/Users/kirill/_my_local/code-local/personal-workspace/flashcards-open-source-app/apps/ios/Flashcards/Flashcards/AI/AIChatStore+RunOrchestration.swift)
 - [apps/ios/Flashcards/Flashcards/AI/AIChatServerErrorSupport.swift](/Users/kirill/_my_local/code-local/personal-workspace/flashcards-open-source-app/apps/ios/Flashcards/Flashcards/AI/AIChatServerErrorSupport.swift)
 - [apps/ios/Flashcards/Flashcards/AI/AIChatVoiceDictation.swift](/Users/kirill/_my_local/code-local/personal-workspace/flashcards-open-source-app/apps/ios/Flashcards/Flashcards/AI/AIChatVoiceDictation.swift)
